pyproject.toml
[tool.poetry]
name = "perry-the-docker-agent"
version = "0.1.0"
description = "Your cool remote docker agent in the cloud"
authors = ["Daverin <mail@daverin.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
boto3 = "1.24.80"
click = "<7.2.0,>=7.1.1"
colorlog = "^6.7.0"
sceptre = "^3.2.0"
pathspec = "^0.10.1"
typer = {version = "^0.6.1", extras = ["all"]}
pydantic = "^1.10.2"
PyYAML = "5.4.1"

[tool.poetry.scripts]
perry = "perry_the_docker_agent:main.app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
